refactors main logic and splits out fetchers and resolvers
