+++ .changeset/pink-bananas-admire.md
@@ -1,0 +1,5 @@
+---
+'@hypermod/cli': minor
+---
+
+Adds the ability to fetch codemods from the hypermod app API. Also includes a large refactor of the source code to accomidate
